// DXC_Frontend/src/api/models/supportGroupCategoryDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface SupportGroupCategoryDto {
  publicId?: string
  /** @nullable */
  name: string | null
  /** @nullable */
  description?: string | null
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
}